requirements.txt
python-chess>=1.10.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
ipywidgets>=8.0.0
zstandard>=0.21.0
pyarrow>=14.0.0
huggingface_hub>=0.19.0
seaborn>=0.12.0
